chore: apply self review concerns
