+++ modules/core/exported/channel.go
@@ -17,23 +17,3 @@ type CounterpartyChannelI interface {
 	GetChannelID() string
 	ValidateBasic() error
 }
-
-// PacketI defines the standard interface for IBC packets
-type PacketI interface {
-	GetSequence() uint64
-	GetTimeoutHeight() Height
-	GetTimeoutTimestamp() uint64
-	GetSourcePort() string
-	GetSourceChannel() string
-	GetDestPort() string
-	GetDestChannel() string
-	GetData() []byte
-	ValidateBasic() error
-}
-
-// Acknowledgement defines the interface used to return
-// acknowledgements in the OnRecvPacket callback.
-type Acknowledgement interface {
-	Success() bool
-	Acknowledgement() []byte
-}
